fix(schematic): remove scrollbars from panel
- Fix bug where scrollbars appear in nested Panel components (prefer
  spacebar panning) [sc-2003]
- Refactor schema builtins for readability (extension of [sc-2001])
- Add down target to base Makefile and README
- Move service conflict section before prepare section for readability
  in README
- Fix spelling error in Panel for panel sizing
- Run vite clean on every app-run make target call to ensure cache is
  gone
