+++ lib/dal/src/schema/builtins.rs
@@ -1,4 +1,5 @@
-use crate::schema::{SchemaResult, SchemaVariant, UiMenu};
+use crate::schema::defaults;
+use crate::schema::{SchemaResult, UiMenu};
 use crate::{
     HistoryActor, Schema, SchemaError, SchemaKind, SchematicKind, StandardModel, Tenancy,
     Visibility,
@@ -32,7 +33,7 @@ async fn application(
     history_actor: &HistoryActor,
 ) -> SchemaResult<()> {
     let name = "application".to_string();
-    let mut schema = match create_schema(
+    let mut schema = match defaults::create_schema(
         txn,
         nats,
         tenancy,
@@ -51,7 +52,7 @@ async fn application(
         .set_ui_hidden(txn, nats, visibility, history_actor, true)
         .await?;
 
-    create_and_set_default_schema_variant(
+    defaults::create_and_set_default_schema_variant(
         txn,
         nats,
         tenancy,
@@ -71,7 +72,7 @@ async fn service(
     history_actor: &HistoryActor,
 ) -> SchemaResult<()> {
     let name = "service".to_string();
-    let mut schema = match create_schema(
+    let mut schema = match defaults::create_schema(
         txn,
         nats,
         tenancy,
@@ -136,7 +137,7 @@ async fn service(
         )
         .await?;
 
-    create_and_set_default_schema_variant(
+    defaults::create_and_set_default_schema_variant(
         txn,
         nats,
         tenancy,
@@ -156,7 +157,7 @@ async fn kubernetes_service(
     history_actor: &HistoryActor,
 ) -> SchemaResult<()> {
     let name = "kubernetes_service".to_string();
-    let mut schema = match create_schema(
+    let mut schema = match defaults::create_schema(
         txn,
         nats,
         tenancy,
@@ -171,7 +172,7 @@ async fn kubernetes_service(
         None => return Ok(()),
     };
 
-    create_and_set_default_schema_variant(
+    defaults::create_and_set_default_schema_variant(
         txn,
         nats,
         tenancy,
@@ -192,7 +193,7 @@ async fn docker_image(
     history_actor: &HistoryActor,
 ) -> SchemaResult<()> {
     let name = "docker_imagee".to_string();
-    let mut schema = match create_schema(
+    let mut schema = match defaults::create_schema(
         txn,
         nats,
         tenancy,
@@ -207,7 +208,7 @@ async fn docker_image(
         None => return Ok(()),
     };
 
-    create_and_set_default_schema_variant(
+    defaults::create_and_set_default_schema_variant(
         txn,
         nats,
         tenancy,
@@ -218,68 +219,4 @@ async fn docker_image(
     .await?;
 
     Ok(())
-}
-
-// TODO(nick): this function should return an "AlreadyExists" error instead of None wrapped by Ok,
-// but since the client will have to deal with same scenario either way, it makes little difference
-// for now.
-async fn create_schema(
-    txn: &PgTxn<'_>,
-    nats: &NatsTxn,
-    tenancy: &Tenancy,
-    visibility: &Visibility,
-    history_actor: &HistoryActor,
-    schema_name: &str,
-    schema_kind: &SchemaKind,
-) -> SchemaResult<Option<Schema>> {
-    match default_schema_exists(txn, tenancy, visibility, schema_name).await? {
-        true => Ok(None),
-        false => {
-            let schema = Schema::new(
-                txn,
-                nats,
-                tenancy,
-                visibility,
-                history_actor,
-                schema_name,
-                schema_kind,
-            )
-            .await?;
-            Ok(Some(schema))
-        }
-    }
-}
-
-// TODO(nick): there's one issue here. If the schema kind has changed, then this check will be
-// inaccurate. As a result, we will be unable to re-create the schema without manual intervention.
-// This should be fine since this code should likely only last as long as default schemas need to
-// be created... which is hopefully not long.... hopefully...
-async fn default_schema_exists(
-    txn: &PgTxn<'_>,
-    tenancy: &Tenancy,
-    visibility: &Visibility,
-    schema_name: &str,
-) -> SchemaResult<bool> {
-    Ok(
-        !Schema::find_by_attr(txn, tenancy, visibility, "name", &schema_name.to_string())
-            .await?
-            .is_empty(),
-    )
-}
-
-async fn create_and_set_default_schema_variant(
-    txn: &PgTxn<'_>,
-    nats: &NatsTxn,
-    tenancy: &Tenancy,
-    visibility: &Visibility,
-    history_actor: &HistoryActor,
-    schema: &mut Schema,
-) -> SchemaResult<()> {
-    let variant = SchemaVariant::new(txn, nats, tenancy, visibility, history_actor, "v0").await?;
-    variant
-        .set_schema(txn, nats, visibility, history_actor, schema.id())
-        .await?;
-    schema
-        .set_default_schema_variant_id(txn, nats, visibility, history_actor, Some(*variant.id()))
-        .await
 }
